Revert "automatically download models from HF (#61)"
This reverts commit ccdf849084dd87e98ee4df876bf89fefa413d124.
