+++ casalioy/startLLM.py
@@ -18,13 +18,12 @@ from casalioy.load_env import (
     model_stop,
     model_temp,
     model_type,
-    n_forward_documents,
     n_gpu_layers,
-    n_retrieve_documents,
     persist_directory,
+    print_HTML,
+    prompt_HTML,
     use_mlock,
 )
-from casalioy.utils import print_HTML, prompt_HTML
 
 
 class QASystem:
@@ -87,7 +86,6 @@ class QASystem:
             return_source_documents=True,
             chain_type_kwargs=get_prompt_template_kwargs(),
         )
-        self.qa.retriever.search_kwargs = {**self.qa.retriever.search_kwargs, "k": n_forward_documents, "fetch_k": n_retrieve_documents}
 
     def prompt_once(self, query: str) -> tuple[str, str]:
         """run a prompt"""
